加入了软件更新和模型更新分开的更新模式，加入了远程注册、心跳和上传图片的功能

+++ updater/updater.cpp
@@ -25,29 +25,58 @@ bool Updater::initFromArgs()
         text += QString("args[%1] = %2\n").arg(i).arg(args[i]);
     }
 
-    QMessageBox::information(nullptr, "Updater Arguments", text);
 
 
 //    args[0]  程序自身路径（永远是 exe）
-//    args[1]  m_zipPath
-//    args[2]  m_targetDir
-//    args[3]  m_mainPid
+//    args[2]  m_FzipPath
+//    args[4]  m_targetFirmDir
+//    args[6]  m_mainPid
+//    args[8]  updateFaA
+    bool ok = false;
 
     for (int i = 0; i < args.size(); ++i)
     {
-        if (args[i] == "--zip" && i + 1 < args.size())
-            m_zipPath = args[i + 1];
+        if (args[i] == "--Fzip" && i + 1 < args.size())
+            m_FzipPath = args[i + 1];
+
+        if (args[i] == "--Azip" && i + 1 < args.size())
+            m_AzipPath = args[i + 1];
 
         if (args[i] == "--target" && i + 1 < args.size())
-            m_targetDir = args[i + 1];
+            m_targetFirmDir = args[i + 1];
 
         if (args[i] == "--pid" && i + 1 < args.size())
             m_mainPid = args[i + 1].toLongLong();
-    }
 
-    return !m_zipPath.isEmpty() &&
-           !m_targetDir.isEmpty() &&
-           m_mainPid > 0;
+        if (args[i] == "--FaA" && i + 1 < args.size())
+        {
+            m_updateFaA = args[i + 1].toLongLong();
+            ok = true;
+        }
+    }
+    if(ok)
+    {
+        if (m_updateFaA == 0) // 软件和算法都不需要更新
+        {
+            return 0;
+        }else if (m_updateFaA == 1) // 仅更新软件
+        {
+            return !m_FzipPath.isEmpty() &&
+                   !m_targetFirmDir.isEmpty() &&
+                   m_mainPid > 0;
+        }else if(m_updateFaA == 2) // 仅更新算法
+        {
+            return !m_AzipPath.isEmpty() &&
+                   !m_targetFirmDir.isEmpty() &&
+                   m_mainPid > 0;
+        }else if(m_updateFaA == 3) // 更新软件和算法
+        {
+            return !m_FzipPath.isEmpty() &&
+                   !m_AzipPath.isEmpty() &&
+                   !m_targetFirmDir.isEmpty() &&
+                   m_mainPid > 0;
+        }
+    }
 }
 
 void Updater::start()
@@ -93,71 +122,148 @@ void Updater::waitForMainProcessExit()
 
 void Updater::extractZip()
 {
-    m_tempDir = m_targetDir + "/update_tmp";
-
-    QDir tempDir(m_tempDir);
-    if (tempDir.exists())
-        tempDir.removeRecursively();
-
-    QDir().mkpath(m_tempDir);
-
-    QProcess process;
-
-#ifdef Q_OS_WIN
-    // Windows 10+ 自带 tar
-    QString program = "tar";
-    QStringList args;
-    args << "-xf" << m_zipPath
-         << "-C" << m_tempDir;
-
-#elif defined(Q_OS_LINUX)
-    QString program = "unzip";
-    QStringList args;
-    args << m_zipPath
-         << "-d" << m_tempDir;
-#endif
-
-    process.start(program, args);
-    process.waitForFinished(-1);
-
-    if (process.exitCode() != 0)
+    if(m_updateFaA & UpdateFirmware)
     {
-        qCritical() << "Extract failed:" << process.readAllStandardError();
+        m_tempFirmDir = m_targetFirmDir + "/update_tmp";  // 这里是存储更新包解压文件的地方
+
+        QDir tempDir(m_tempFirmDir);
+        if (tempDir.exists())
+            tempDir.removeRecursively();
+
+        QDir().mkpath(m_tempFirmDir);
+
+        QProcess process;
+
+    #ifdef Q_OS_WIN
+        // Windows 10+ 自带 tar
+        QString program = "tar";
+        QStringList args;
+        args << "-xf" << m_FzipPath
+             << "-C" << m_tempFirmDir;
+
+    #elif defined(Q_OS_LINUX)
+        QString program = "unzip";
+        QStringList args;
+        args << m_FzipPath
+             << "-d" << m_tempFirmDir;
+    #endif
+
+        process.start(program, args);
+        process.waitForFinished(-1);
+
+        if (process.exitCode() != 0)
+        {
+            qCritical() << "Extract failed:" << process.readAllStandardError();
+        }
+    }
+    if(m_updateFaA & UpdateAlgorithm)
+    {
+
+        m_tempAlgorithmDir = m_targetAlgorithmDir + "/Algorithm_update_tmp";
+
+        QDir tempDir(m_tempAlgorithmDir);
+        if (tempDir.exists())
+            tempDir.removeRecursively();
+
+        QDir().mkpath(m_tempAlgorithmDir);
+
+        QProcess process;
+
+    #ifdef Q_OS_WIN
+        // Windows 10+ 自带 tar
+        QString program = "tar";
+        QStringList args;
+        args << "-xf" << m_AzipPath
+             << "-C" << m_tempAlgorithmDir;
+
+    #elif defined(Q_OS_LINUX)
+        QString program = "unzip";
+        QStringList args;
+        args << m_AzipPath
+             << "-d" << m_tempAlgorithmDir;
+    #endif
+
+        process.start(program, args);
+        process.waitForFinished(-1);
+
+        if (process.exitCode() != 0)
+        {
+            qCritical() << "Extract failed:" << process.readAllStandardError();
+        }
     }
 }
 void Updater::backupOldVersion()
 {
-    QString backupZip = m_targetDir + "/pastVersion.zip";
+    if (m_updateFaA & UpdateFirmware)   // 更新软件
+    {
+        QString backupZip = m_targetFirmDir + "/pastFirmVersion.zip";
 
-    if (QFile::exists(backupZip))
-        QFile::remove(backupZip);
+        if (QFile::exists(backupZip))
+            QFile::remove(backupZip);
 
-    QProcess process;
+        QProcess process;
 
-#ifdef Q_OS_WIN
-    // Windows 使用 tar
-    QString program = "tar";
-    QStringList args;
+    #ifdef Q_OS_WIN
+        // Windows 使用 tar
+        QString program = "tar";
+        QStringList args;
 
-    args << "-cf" << backupZip
-         << "guangxuan.exe"
-         << "guangxuan_zh_CN.qm"
-         << "guangxuan_zh_EN.qm";
+        args << "-cf" << backupZip
+             << "guangxuan.exe"
+             << "guangxuan_zh_CN.qm"
+             << "guangxuan_zh_EN.qm";
 
-#elif defined(Q_OS_LINUX)
-    // Linux 使用 zip
-    QString program = "zip";
-    QStringList args;
+    #elif defined(Q_OS_LINUX)
+        // Linux 使用 zip
+        QString program = "zip";
+        QStringList args;
 
-    args << "-r" << backupZip
-         << "guangxuan"
-         << "guangxuan_zh_CN.qm"
-         << "guangxuan_zh_EN.qm";
-#endif
+        args << "-r" << backupZip
+             << "guangxuan"
+             << "guangxuan_zh_CN.qm"
+             << "guangxuan_zh_EN.qm";
+    #endif
 
-    process.setWorkingDirectory(m_targetDir);
-    process.start(program, args);
-    process.waitForFinished(-1);
+        process.setWorkingDirectory(m_targetFirmDir);
+        process.start(program, args);
+        process.waitForFinished(-1);
+    }
+
+    if(m_updateFaA & UpdateAlgorithm)
+    {
+
+        QString backupZip = m_targetAlgorithmDir + "/pastAlgorithmVersion.zip";
+
+        if (QFile::exists(backupZip))
+            QFile::remove(backupZip);
+
+        QProcess process;
+
+    // TODO 这下面要改成算法专用的更新
+    #ifdef Q_OS_WIN
+        QString program = "tar";
+        QStringList args;
+
+        args << "-cf" << backupZip
+             << "guangxuan.exe"
+             << "guangxuan_zh_CN.qm"
+             << "guangxuan_zh_EN.qm";
+
+    #elif defined(Q_OS_LINUX)
+        // Linux 使用 zip
+        QString program = "zip";
+        QStringList args;
+
+        args << "-r" << backupZip
+             << "guangxuan"
+             << "guangxuan_zh_CN.qm"
+             << "guangxuan_zh_EN.qm";
+    #endif
+
+        process.setWorkingDirectory(m_targetAlgorithmDir);
+        process.start(program, args);
+        process.waitForFinished(-1);
+    }
 }
 
 bool Updater::copyRecursively(const QString& src, const QString& dst)
@@ -193,26 +299,55 @@ void Updater::replaceFiles()
 {
     backupOldVersion();
 
-    QDir srcDir(m_tempDir);
-    QDir dstDir(m_targetDir);
-
-    QFileInfoList entries = srcDir.entryInfoList(
-        QDir::NoDotAndDotDot | QDir::AllEntries);
-
-    for (const QFileInfo& entry : entries)
+    if(m_updateFaA & UpdateFirmware)
     {
-        QString srcPath = entry.absoluteFilePath();
-        QString dstPath = dstDir.filePath(entry.fileName());
+        QDir srcDir(m_tempFirmDir);
+        QDir dstDir(m_targetFirmDir);
 
-        if (entry.isDir())
+        QFileInfoList entries = srcDir.entryInfoList(
+            QDir::NoDotAndDotDot | QDir::AllEntries);
+
+        for (const QFileInfo& entry : entries)
         {
-            QDir().mkpath(dstPath);
-            copyRecursively(srcPath, dstPath);
+            QString srcPath = entry.absoluteFilePath();
+            QString dstPath = dstDir.filePath(entry.fileName());
+
+            if (entry.isDir())
+            {
+                QDir().mkpath(dstPath);
+                copyRecursively(srcPath, dstPath);
+            }
+            else
+            {
+                QFile::remove(dstPath);
+                QFile::copy(srcPath, dstPath);
+            }
         }
-        else
+    }
+
+    if(m_updateFaA & UpdateAlgorithm)
+    {
+        QDir srcDir(m_tempAlgorithmDir);
+        QDir dstDir(m_targetAlgorithmDir);
+
+        QFileInfoList entries = srcDir.entryInfoList(
+            QDir::NoDotAndDotDot | QDir::AllEntries);
+
+        for (const QFileInfo& entry : entries)
         {
-            QFile::remove(dstPath);
-            QFile::copy(srcPath, dstPath);
+            QString srcPath = entry.absoluteFilePath();
+            QString dstPath = dstDir.filePath(entry.fileName());
+
+            if (entry.isDir())
+            {
+                QDir().mkpath(dstPath);
+                copyRecursively(srcPath, dstPath);
+            }
+            else
+            {
+                QFile::remove(dstPath);
+                QFile::copy(srcPath, dstPath);
+            }
         }
     }
 }
@@ -220,5 +355,13 @@ void Updater::replaceFiles()
 
 void Updater::restartMainApp()
 {
-    QProcess::startDetached(m_targetDir + "/guangxuan.exe");
+    if (m_updateFaA & UpdateFirmware)
+    {
+        QProcess::startDetached(m_targetFirmDir + "/guangxuan.exe");
+    }
+
+    if (m_updateFaA & UpdateAlgorithm)
+    {
+        QProcess::startDetached(m_targetAlgorithmDir + "/guangxuan.exe");  // 更换算法模型的路径
+    }
 }
